requirements.txt
streamlit==1.39.0
google-cloud-storage==2.18.2
google-cloud-bigquery==3.25.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1